requirements.txt
# PDF版面分析依赖包
layoutparser>=0.3.4
pdf2image>=3.1.0
opencv-python>=4.8.0
numpy>=1.21.0
Pillow>=9.0.0
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.21.0
detectron2>=0.6
# 可选：如果需要使用PaddleOCR
# paddlepaddle>=2.4.0
# paddleocr>=2.7.0
# 可选：如果需要使用PyMuPDF
# PyMuPDF>=1.20.0

